+++ multitld/include/detector/nn/NNScore.hpp
@@ -21,6 +21,6 @@ public:
     Score* sum(Score* other);
     Score* divide(int n);
     Score* clone();
-    float getValue(int modelId);
+    double getValue(int modelId);
 };
 #endif
